requirements.txt
flask>=3.0.0
streamlit>=1.34.0
ultralytics>=8.2.0
opencv-python>=4.9.0
paho-mqtt>=1.6.1
numpy>=1.26.0
pandas>=2.2.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.1
